create new files for front propagation and save the files in the proper locations
