+++ ImageInLib/ImageInLib/ImageInLib.vcxproj.filters
@@ -153,6 +153,18 @@
     <Filter Include="Source Files\morphology">
       <UniqueIdentifier>{4391e895-a529-434c-b679-ffc255f9b1ae}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Source Files\Front Popagation">
+      <UniqueIdentifier>{36ab1fcf-3a04-40a0-9417-e3b735f1efc6}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Source Files\Interfaces\Front Propagation">
+      <UniqueIdentifier>{c5469f28-b256-4cb7-9e80-79f612705d0f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\Front propagation">
+      <UniqueIdentifier>{30a00689-394c-48d7-b5ef-e72dd8b4ee6f}</UniqueIdentifier>
+    </Filter>
+    <Filter Include="Header Files\Interfaces\Front propagation">
+      <UniqueIdentifier>{2e9c3afb-187d-4f5b-96ff-c2396c3fd5d2}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\transformations.c">
@@ -338,6 +350,9 @@
     <ClCompile Include="src\generate_3d_curves.c">
       <Filter>Source Files\Shape Generation</Filter>
     </ClCompile>
+    <ClCompile Include="src\front_propagation.cpp">
+      <Filter>Source Files\Front Popagation</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\transformations.h">
@@ -513,6 +528,9 @@
     </ClInclude>
     <ClInclude Include="src\generate_3d_curves.h">
       <Filter>Header Files\Shape Generation</Filter>
+    </ClInclude>
+    <ClInclude Include="include\front_propagation.h">
+      <Filter>Header Files\Front propagation</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
